BugFix Fix problems with cpp2py_converter_table
We no longer fetch the table directly from the __main__ module.
This was leading to trouble due to the order in which modules
were unloaded in the Py_FinalizeEx.

Instead, each module has a shared_ptr to the common map which
in the module initialization is set as follows

1. Check if the __main__ module has an atribute __cpp2py_table
containing a ptr to a shared_ptr<map> on the head. If so use it
to init your own
2. If not, we initialize our own shared ptr with a new map.
   We also create an additional shared ptr on the heap which we
   attach to __main__ as __cpp2py_table. We also register the
   proper destruction function for this second shared ptr.